SNAKE_GAME: refactored

+++ src/renderer.cpp
@@ -10,25 +10,21 @@ Renderer::Renderer(const std::size_t& screen_width,
       grid_width(grid_width),
       grid_height(grid_height)
 {
-    // Initialize SDL
     if (SDL_Init(SDL_INIT_VIDEO) < 0)
     {
         std::cerr << "SDL could not initialize.\n";
         std::cerr << "SDL_Error: " << SDL_GetError() << "\n";
     }
 
-    // Create Window
     sdl_window = SDL_CreateWindow("Snake Game", SDL_WINDOWPOS_CENTERED,
-                                  SDL_WINDOWPOS_CENTERED, screen_width,
-                                  screen_height, SDL_WINDOW_SHOWN);
-
+                                  SDL_WINDOWPOS_CENTERED, static_cast<int>(screen_width),
+                                  static_cast<int>(screen_height), SDL_WINDOW_SHOWN);
     if (nullptr == sdl_window)
     {
         std::cerr << "Window could not be created.\n";
         std::cerr << " SDL_Error: " << SDL_GetError() << "\n";
     }
 
-    // Create renderer
     sdl_renderer = SDL_CreateRenderer(sdl_window, -1, SDL_RENDERER_SOFTWARE);
     if (nullptr == sdl_renderer)
     {
@@ -49,17 +45,14 @@ void Renderer::Render(Snake const& snake, Snake const& auto_snake, SDL_Point con
     block.w = screen_width / grid_width;
     block.h = screen_height / grid_height;
 
-    // Clear screen
     SDL_SetRenderDrawColor(sdl_renderer, 0x1E, 0x1E, 0x1E, 0xFF);
     SDL_RenderClear(sdl_renderer);
 
-    // Render food
     SDL_SetRenderDrawColor(sdl_renderer, 0xFF, 0xCC, 0x00, 0xFF);
     block.x = food.x * block.w;
     block.y = food.y * block.h;
     SDL_RenderFillRect(sdl_renderer, &block);
 
-    // Render snake_1's body
     SDL_SetRenderDrawColor(sdl_renderer, 0xFF, 0xFF, 0xFF, 0xFF);
     for (SDL_Point const& point : snake.body)
     {
@@ -68,7 +61,6 @@ void Renderer::Render(Snake const& snake, Snake const& auto_snake, SDL_Point con
         SDL_RenderFillRect(sdl_renderer, &block);
     }
 
-    // Render snake_2's body
     SDL_SetRenderDrawColor(sdl_renderer, 0x8F, 0xAA, 0xDC, 0xFF);
     for (SDL_Point const& point : auto_snake.body)
     {
@@ -77,7 +69,6 @@ void Renderer::Render(Snake const& snake, Snake const& auto_snake, SDL_Point con
         SDL_RenderFillRect(sdl_renderer, &block);
     }
 
-    // Render snake 1's head
     block.x = static_cast<int>(snake.head_x) * block.w;
     block.y = static_cast<int>(snake.head_y) * block.h;
     if (snake.alive)
@@ -90,7 +81,6 @@ void Renderer::Render(Snake const& snake, Snake const& auto_snake, SDL_Point con
     }
     SDL_RenderFillRect(sdl_renderer, &block);
 
-    // Render snake 2's head
     block.x = static_cast<int>(auto_snake.head_x) * block.w;
     block.y = static_cast<int>(auto_snake.head_y) * block.h;
     if (auto_snake.alive)
@@ -103,7 +93,6 @@ void Renderer::Render(Snake const& snake, Snake const& auto_snake, SDL_Point con
     }
     SDL_RenderFillRect(sdl_renderer, &block);
 
-    // Update Screen
     SDL_RenderPresent(sdl_renderer);
 }
